requirements.txt
asgiref==3.3.4
deap==1.3.1
dj-database-url==0.5.0
Django==3.2.3
gunicorn==20.1.0
haversine==2.3.1
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
django-leaflet==0.29.0
django-geojson==3.3.0
numpy==1.26.0
pandas==2.1.1
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
django-leaflet==0.29.0
django-geojson==3.3.0
numpy==1.26.0
pandas==2.1.1
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==4.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.0
boto3==1.28.44